backend/requirements.txt
asgiref==3.10.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cssselect2==0.8.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.60.1
idna==3.11
pillow==12.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dotenv==1.2.1
qrcode==8.2
reportlab==4.4.4
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
stripe==7.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
wheel==0.45.1
zopfli==0.2.3.post1
reportlab==4.0.7